requirements.txt
google-api-python-client==2.97.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.1.0
httplib2==0.22.0
